% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{AverageDownTradesAnalysisBlockClient-class}
\alias{AverageDownTradesAnalysisBlockClient-class}
\title{Clas wrapping access to AverageDownTradesAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for AverageDownTradesAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

