% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{AverageDownTradesFocusAnalysisBlockClient-class}
\alias{AverageDownTradesFocusAnalysisBlockClient-class}
\title{Clas wrapping access to AverageDownTradesFocusAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for AverageDownTradesFocusAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

