% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{BuysAndSellsAnalysisBlockClient-class}
\alias{BuysAndSellsAnalysisBlockClient-class}
\title{Clas wrapping access to BuysAndSellsAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for BuysAndSellsAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

