% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_client.r
\docType{methods}
\name{dataRequest,VirtualAnalysisObjectstoreClient,data.frame-method}
\alias{dataRequest,VirtualAnalysisObjectstoreClient,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{VirtualAnalysisObjectstoreClient,data.frame}(object,
  key_values, force = FALSE)
}
\arguments{
\item{object}{object of class 'VirtualAnalysisObjectstoreClient'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'VirtualAnalysisObjectstoreClient'.
}
\description{
Generic method to request data from data source.
Needs to be implemented in derived classes to work
}

