% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.r
\docType{class}
\name{Engine-class}
\alias{Engine-class}
\title{Class inmplementing Frontend Engine Server}
\description{
Implements Server opening port and accepting commands
from client via socket
}
\section{Slots}{

\describe{
\item{\code{interpreter}}{"EngineCommandInterpreter",}

\item{\code{socket}}{"ProcessSocket",}

\item{\code{store_name}}{"character",}

\item{\code{analys_str}}{"AnalysisObjectStore",}

\item{\code{app_store}}{"BasicAppClient",}

\item{\code{trader}}{"numeric",}

\item{\code{module_date}}{"character",}

\item{\code{lookback}}{"character",}

\item{\code{key_hash}}{"character",}

\item{\code{module_name}}{"character",}

\item{\code{app_path}}{"character",}

\item{\code{app_port}}{"character",}

\item{\code{app_host}}{"character",}

\item{\code{response}}{"character",}

\item{\code{module}}{"list",}

\item{\code{app}}{"function",}

\item{\code{analysis_ggplot}}{"gg",}

\item{\code{analysis_data}}{"data.frame"}
}}

