% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.r
\docType{class}
\name{EngineCommandInterpreter-class}
\alias{EngineCommandInterpreter-class}
\title{Class inmplementing Frontend Engine Command Interpreter}
\description{
Implements Server opening port and accepting commands
from client via socket
}
\section{Slots}{

\describe{
\item{\code{snd_respnse}}{"logical"}

\item{\code{snd_data}}{"logical"}

\item{\code{run_app}}{"logical"}

\item{\code{callback}}{"function"}

\item{\code{split_char}}{"character"}

\item{\code{cmmd_lst}}{"character"}
}}

