% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{ExtendedTradesAnalysisBlockClient-class}
\alias{ExtendedTradesAnalysisBlockClient-class}
\title{Clas wrapping access to ExtendedTradesAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for ExtendedTradesAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

