% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_client.r
\docType{methods}
\name{getAnalysisBlock}
\alias{getAnalysisBlock}
\alias{getAnalysisBlock,VirtualAnalysisObjectstoreClient-method}
\title{Returns contained analysis block object}
\usage{
getAnalysisBlock(object)

\S4method{getAnalysisBlock}{VirtualAnalysisObjectstoreClient}(object)
}
\arguments{
\item{object}{: object of type "VirtualAnalysisObjectstoreClient"}
}
\value{
\code{analysis_block} object of class "VirtualAnalysisBlock"
}
\description{
Returns contained analysis block object
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAnalysisObjectstoreClient}: Returns contained analysis block object
}}

