% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_client.r
\docType{methods}
\name{getAnalysisClass}
\alias{getAnalysisClass}
\alias{getAnalysisClass,VirtualAnalysisObjectstoreClient-method}
\title{Returns name of the underlying analysis block class.}
\usage{
getAnalysisClass(object)

\S4method{getAnalysisClass}{VirtualAnalysisObjectstoreClient}(object)
}
\arguments{
\item{object}{: object of type "VirtualAnalysisObjectstoreClient"}
}
\value{
\code{analysis_class} "character" name of the analysis class
}
\description{
Returns name of the underlying analysis block class.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAnalysisObjectstoreClient}: Returns name of the underlying analysis block class.
}}

