% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_factories.r
\docType{methods}
\name{getServer}
\alias{getServer}
\alias{getServer,ShinyFactory-method}
\title{Get stored Server Object}
\usage{
getServer(object, analysis_ggplot, analysis_data)

\S4method{getServer}{ShinyFactory}(object)
}
\arguments{
\item{object}{object of class "ShinyFactory"}

\item{analysis_ggplot}{"ggplot" object}

\item{analysis_data}{"data.frame" with data of ggplot}
}
\value{
\code{ui} shiny UI function
}
\description{
Get stored Server Object
}
\section{Methods (by class)}{
\itemize{
\item \code{ShinyFactory}: Get stored UI Object
}}

