% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{IndexPortfolioFactorExposuresAnalysisBlockClient-class}
\alias{IndexPortfolioFactorExposuresAnalysisBlockClient-class}
\title{Clas wrapping access to IndexPortfolioFactorExposuresAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for IndexPortfolioFactorExposuresAnalysisBlock class
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

