% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{IndexPortfolioFactorReturnsAnalysisBlockClient-class}
\alias{IndexPortfolioFactorReturnsAnalysisBlockClient-class}
\title{Clas wrapping access to IndexPortfolioFactorReturnsAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for IndexPortfolioFactorReturnsAnalysisBlock class
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

