% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{IndexPortfolioVarianceDecompositionAnalysisBlockClient-class}
\alias{IndexPortfolioVarianceDecompositionAnalysisBlockClient-class}
\title{Clas wrapping access to IndexPortfolioVarianceDecompositionAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for IndexPortfolioVarianceDecompositionAnalysisBlock class
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

