% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{MarketReturnAnalysisBlockClient-class}
\alias{MarketReturnAnalysisBlockClient-class}
\title{Clas wrapping access to MarketReturnAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for MarketReturnAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

