% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{MarketStyleFactorStatisticAnalysisBlockClient-class}
\alias{MarketStyleFactorStatisticAnalysisBlockClient-class}
\title{Clas wrapping access to MarketStyleFactorStatisticAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for MarketStyleFactorStatisticAnalysisBlock class
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

