% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{OffsidePositionsAnalysisBlockClient-class}
\alias{OffsidePositionsAnalysisBlockClient-class}
\title{Clas wrapping access to OffsidePositionsAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for OffsidePositionsAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

