% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{OffsidePositionsCumulativePnLAnalysisBlockClient-class}
\alias{OffsidePositionsCumulativePnLAnalysisBlockClient-class}
\title{Clas wrapping access to OffsidePositionsCumulativePnLAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for OffsidePositionsCumulativePnLAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

