% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{PnLTradedInLongShortHedgeAnalysisBlockClient-class}
\alias{PnLTradedInLongShortHedgeAnalysisBlockClient-class}
\title{Clas wrapping access to PnLTradedInLongShortHedgeAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for PnLTradedInLongShortHedgeAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

