% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{PortfolioFactorExposureAnalysisBlockClient-class}
\alias{PortfolioFactorExposureAnalysisBlockClient-class}
\title{Clas wrapping access to PortfolioFactorExposureAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for PortfolioFactorExposureAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

