% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{PortfolioFactorExposuresAnalysisBlockClient-class}
\alias{PortfolioFactorExposuresAnalysisBlockClient-class}
\title{Clas wrapping access to PortfolioFactorExposuresAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for PortfolioFactorExposuresAnalysisBlock class
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

