% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{PortfolioFactorReturnsAnalysisBlockClient-class}
\alias{PortfolioFactorReturnsAnalysisBlockClient-class}
\title{Clas wrapping access to PortfolioFactorReturnsAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for PortfolioFactorReturnsAnalysisBlock class
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

