% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{PortfolioVarianceDecompositionAnalysisBlockClient-class}
\alias{PortfolioVarianceDecompositionAnalysisBlockClient-class}
\title{Clas wrapping access to PortfolioVarianceDecompositionAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for PortfolioVarianceDecompositionAnalysisBlock class
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

