% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{PositionRevisitsAnalysisBlockClient-class}
\alias{PositionRevisitsAnalysisBlockClient-class}
\title{Clas wrapping access to PositionRevisitsAnalysisBlockClient}
\description{
Implements methods to access analysis objectstore
for PositionRevisitsAnalysisBlockClient class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

