% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{PositionsHoldingPeriodAnalysisBlockClient-class}
\alias{PositionsHoldingPeriodAnalysisBlockClient-class}
\title{Clas wrapping access to PositionsHoldingPeriodAnalysisBlockClient}
\description{
Implements methods to access analysis objectstore
for PositionsHoldingPeriodAnalysisBlock class
}
\details{
Inherits from: "VirtualStrategyPortfolioAnalysisBlockClient"
}

