% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_factories.r
\docType{methods}
\name{shinyServerFactory}
\alias{shinyServerFactory}
\alias{shinyServerFactory,ShinyFactory-method}
\title{generate Server Object}
\usage{
shinyServerFactory(object, analysis_ggplot, analysis_data,
  ui_options = list())

\S4method{shinyServerFactory}{ShinyFactory}(object, analysis_ggplot,
  analysis_data, ui_options = list())
}
\arguments{
\item{object}{object of class "ShinyFactory"}

\item{analysis_ggplot}{"ggplot" object}

\item{analysis_data}{"data.frame" ggplot data}

\item{ui_options}{"list" with additional UI options}
}
\value{
\code{object} object of class "ShinyFactory"
}
\description{
generate Server Object
}
\section{Methods (by class)}{
\itemize{
\item \code{ShinyFactory}: generate Server Object
}}

