% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_factories.r
\docType{methods}
\name{shinyUIFactory}
\alias{shinyUIFactory}
\alias{shinyUIFactory,ShinyFactory-method}
\title{generate UI Object}
\usage{
shinyUIFactory(object, analysis_ggplot, analysis_data, ui_options = list(omit
  = c("Value")))

\S4method{shinyUIFactory}{ShinyFactory}(object, analysis_ggplot, analysis_data,
  ui_options = list(omit = c("Value")))
}
\arguments{
\item{object}{object of class "ShinyFactory"}

\item{analysis_ggplot}{"ggplot" object}

\item{analysis_data}{"data.frame" ggplot data}

\item{ui_options}{"list" with additional UI options}
}
\value{
\code{object} object of class "ShinyFactory"
}
\description{
generate UI Object
}
\section{Methods (by class)}{
\itemize{
\item \code{ShinyFactory}: generate UI Object
}}

