% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.r
\docType{methods}
\name{startEngine}
\alias{startEngine}
\alias{startEngine,Engine-method}
\title{Start Frontend Engine Process}
\usage{
startEngine(object)

\S4method{startEngine}{Engine}(object)
}
\arguments{
\item{object}{object of class "Engine"}
}
\description{
Start Frontend Engine Process
}
\section{Methods (by class)}{
\itemize{
\item \code{Engine}: Start Frontend Engine
}}

