% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{StrategyBreakdownAnalysisBlockClient-class}
\alias{StrategyBreakdownAnalysisBlockClient-class}
\title{Clas wrapping access to StrategyBreakdownAnalysisBlock}
\description{
Implements methods to access analysis objectstore
for StrategyBreakdownAnalysisBlock class
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

