% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.FrontendEngine.r
\docType{package}
\name{TE.FrontendEngine}
\alias{TE.FrontendEngine}
\alias{TE.FrontendEngine-package}
\title{TE.FrontendEngine: Trading Enhancement engine for frontend visualisation}
\description{
The TE.FrontendEngine package classes and functions implementing
Visualization of processing data via shiny applications that
can be generated by server process upon request from client to the
Engine
}

