% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_client.r
\docType{class}
\name{VirtualAnalysisObjectstoreClient-class}
\alias{VirtualAnalysisObjectstoreClient-class}
\title{Clas wrapping access to Analysis block objectstore}
\description{
Implements methods to access analysis objectstore
for specific Analysis classes
}
\details{
Inherits from : "VirtualDataSourceClient"
}
\section{Slots}{

\describe{
\item{\code{analysis_class}}{"character" name of the class of the analyzer}

\item{\code{analysis_block}}{object of class "VirtualAnalysisBlock"}
}}

