% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{VirtualIndexPortfolioAnalysisBlockClient-class}
\alias{VirtualIndexPortfolioAnalysisBlockClient-class}
\title{Index Portfolio classes}
\description{
Implements methods to access analysis objectstore
for IndexPortfolio Risk Analysis
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

