% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_library.r
\docType{class}
\name{VirtualStrategyPortfolioAnalysisBlockClient-class}
\alias{VirtualStrategyPortfolioAnalysisBlockClient-class}
\title{Strategy Portfolio classes}
\description{
Implements methods to access analysis objectstore
for StrategyPortfolio Risk Analysis
}
\details{
Inherits from: "VirtualAnalysisObjectstoreClient"
}

